Refactor backend API responses to use ApiResult wrapper for consistency and clarity. Update CORS configuration to restrict allowed headers. Change PostgreSQL dependency from runtimeOnly to implementation in build.gradle. Enhance exception handling with new ForbiddenException and improve error messages. Update various controllers and services to return ApiResult types, ensuring a unified response structure across the application.

+++ backend/src/main/resources/db/migration/V1__initial_schema.sql
@@ -33,8 +33,8 @@ CREATE TABLE authorities (
 -- -----------------------------------------------------------------------------
 CREATE TABLE carts (
     id          BIGSERIAL PRIMARY KEY NOT NULL,
-    customer_id INTEGER UNIQUE        NOT NULL,
-    total_price NUMERIC(19,4)       NOT NULL DEFAULT 0,
+    customer_id BIGINT UNIQUE        NOT NULL,
+    total_price NUMERIC(19,4)    NOT NULL DEFAULT 0,
     CONSTRAINT fk_cart_customer FOREIGN KEY (customer_id) REFERENCES customers (id) ON DELETE CASCADE
 );
